Configure for Android
OpenList from Planner changed from list type to TArray.

+++ Plugins/GOAP/Source/GOAP/Public/GOAPNode.h
@@ -26,7 +26,7 @@ private:
 
 	int h;
 
-	int g;
+	float g;
 
 	int parent;
